backend/openfaas/auth-user/requirements.txt
pyqrcode
pypng
psycopg2-binary
cryptography
pyotp
bcrypt